Adds all new product/attribute pairs for single attribute not yet in SBA
Addresses the following SBA forum post:

This only evaluates the SBA table looking for the product and a match of
a single option value. If that is not found for the current product, then
the product with its single option value is added to the database as a
variant.

The implemented design of SBA (concatenating attribute indices as text)
complicatess trying to add combinations through this "simple" query.

As demonstrated throughout the SBA code, attribute combinations require
either additional PHP coding beyond simple look ups or complicated queries.